feat: 增加乐观锁字段，数据库时间字段增加时区

+++ src/main/java/wander/nights/forma/event/outbox/OutboxEvent.java
@@ -80,7 +80,7 @@ public class OutboxEvent {
 
     @Version
     @Column(name = "version")
-    private long version = 0;
+    private int version = 0;
 
     public enum Status {
         PENDING, SENT, FAILED
